Changed folder struct & language namespace. language to LANGUAGE, Ilanguage to ILanguage, atom.ts move to project_common.ts
